+++ packages/utils/plugin/sdk/index.ts
@@ -20,6 +20,7 @@ export * from './flow'
 export * from './hooks/index'
 export * from './intelligence'
 
+export * from './quick-actions-sdk'
 export * from './meta-sdk'
 export * from './notification'
 export * from './performance'